+++ lib/redox/models/message.rb
@@ -5,6 +5,7 @@ require_relative "meta"
 module Redox
   module Models
     class Message < Redox::Model
+      attr_accessor :_body
       redox_property :Meta, coerce: Redox::Models::Meta
     end
   end