+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## [Unreleased]
 ### Added
+- _body attribute containing raw response text on query results
 
 ### Changed